Updated Xcode project
